Fix clippy lints

+++ src/bytecode.rs
@@ -233,8 +233,8 @@ impl Iterator for BlockIterator<'_> {
             OpCode::Pop => Op::Pop,
             OpCode::BindLocal => Op::BindLocal(self.next_word() as usize),
             OpCode::PushLocal => Op::PushLocal(self.next_word() as usize),
-            OpCode::BindProp => Op::BindProp(PropId(self.next_word() as u64)),
-            OpCode::PushProp => Op::PushProp(PropId(self.next_word() as u64)),
+            OpCode::BindProp => Op::BindProp(PropId(self.next_word())),
+            OpCode::PushProp => Op::PushProp(PropId(self.next_word())),
             OpCode::Add => Op::Add,
             OpCode::Sub => Op::Sub,
             OpCode::Mul => Op::Mul,
@@ -249,8 +249,8 @@ impl Iterator for BlockIterator<'_> {
             OpCode::Or => Op::Or,
             OpCode::Not => Op::Not,
             OpCode::Print => Op::Print,
-            OpCode::ObjectId => Op::ObjectId(ObjectId::new(self.next_word() as u64)),
-            OpCode::Malloc => Op::Malloc(ObjectId::new(self.next_word() as u64)),
+            OpCode::ObjectId => Op::ObjectId(ObjectId::new(self.next_word())),
+            OpCode::Malloc => Op::Malloc(ObjectId::new(self.next_word())),
         };
         Some(op)
     }
